salsa20-core: Move all code into `lib.rs`
There's practically nothing in `lib.rs` right now aside from extern
crate directives and the `mod`/`use` directives for `SalsaFamilyState`,
so I think it makes sense to just consolidate the crate into a single
file.
